adjusting errors with tripoulies's tester

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: dbessa <dbessa@student.42.rio>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/19 11:08:39 by dbessa            #+#    #+#             */
-/*   Updated: 2023/10/22 19:35:28 by dbessa           ###   ########.fr       */
+/*   Updated: 2023/10/23 15:32:54 by dbessa           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ int		ft_putchar(char c);
 int		ft_putnbr(int nb);
 int		ft_putstr(char *str);
 int		ft_putunsnbr(unsigned int nb);
-int		ft_puthexa(unsigned int d, int arg);
+int		ft_puthexa(unsigned long d, int arg);
 int		check_argument(const char *arg, int i, va_list ap);
 
 #endif
